+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Transition timestamp.
+
+### Fixed
+
 - PSP RBAC.
 
 ## [0.1.3] - 2024-08-13